refactor: move mutations to module
